Consolidate reporter calls into the metric handler.
